+++ myqueue/local.py
@@ -17,6 +17,7 @@ class LocalSchedulerError(Exception):
 class LocalScheduler(Scheduler):
     def submit(self, task: Task, dry_run: bool = False) -> None:
         assert not dry_run
+        task.cmd.function = None
         (id,) = self.send('submit', task)
         task.id = id